临时修改

+++ .eslintrc.js
@@ -26,7 +26,7 @@ module.exports = {
     // allow debugger during development
     'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     'semi': ['error','always'],
-    "quotes":['error', 'double'],
+    "quotes":['error', 'single'],
     "space-before-function-paren":0,
     "no-unused-vars":0,
     "indent":0,
